+++ exporters/otlp/profiles/src/main/java/io/opentelemetry/exporter/otlp/profiles/LineData.java
@@ -10,13 +10,13 @@ import javax.annotation.concurrent.Immutable;
 /**
  * Details a specific line in a source code, linked to a function.
  *
- * @see "pprofextended.proto::Line"
+ * @see "profiles.proto::Line"
  */
 @Immutable
 public interface LineData {
 
   /** The index of the corresponding Function for this line. Index into function table. */
-  long getFunctionIndex();
+  int getFunctionIndex();
 
   /** Line number in source code. */
   long getLine();
